Sparc64: partially revert r515: long long is safer than int64_t
Host headers may define int64_t as different type than long long, especially
on a 64 bit host. This matters with printf formatting.


git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@555 f158a5a8-5612-0410-a976-696ce0be7e32

+++ include/sparc64/types.h
@@ -29,8 +29,8 @@ typedef long long       int64_t;
 #include "autoconf.h"
 
 /* cell based types */
-typedef int64_t   cell;
-typedef uint64_t  ucell;
+typedef long long          cell;
+typedef unsigned long long ucell;
 
 #define FMT_cell    "%lld"
 #define FMT_ucell   "%llu"
